Move enum helper interfaces to upper level
and move them to uk.org.squeezer.framework
