+++ Squeezer/src/main/java/uk/org/ngo/squeezer/framework/EnumWithText.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2015 Kurt Aaholst <kaaholst@gmail.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.org.ngo.squeezer.framework;
+
+import android.content.Context;
+
+/**
+ * Helper interface to associate a text suitable for display in the UI with an enum value.
+ */
+public interface EnumWithText {
+    String getText(Context context);
+}
