+++ Squeezer/src/main/java/uk/org/ngo/squeezer/menu/ViewMenuItemFragment.java
@@ -7,9 +7,10 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 
 import uk.org.ngo.squeezer.R;
+import uk.org.ngo.squeezer.framework.EnumWithText;
+import uk.org.ngo.squeezer.framework.EnumWithTextAndIcon;
 import uk.org.ngo.squeezer.framework.Item;
 import uk.org.ngo.squeezer.framework.ItemAdapter;
-import uk.org.ngo.squeezer.itemlist.dialog.BaseViewDialog;
 
 /**
  * A fragment that implements a "View" menu.
@@ -58,8 +59,8 @@ public class ViewMenuItemFragment extends BaseMenuFragment {
      * Interface that activities that host this fragment must implement.
      */
     public interface ListActivityWithViewMenu<T extends Item,
-            ListLayout extends Enum<ListLayout> & BaseViewDialog.EnumWithTextAndIcon,
-    SortOrder extends Enum<SortOrder> & BaseViewDialog.EnumWithText> {
+            ListLayout extends Enum<ListLayout> & EnumWithTextAndIcon,
+    SortOrder extends Enum<SortOrder> & EnumWithText> {
 
         /**
          * Show a dialog allowing the user to choose the sort order.
